refactor(webapp): move device and guide stores to runes
